requirements.txt
Pillow
numpy
gradio
loguru
scipy
opencv-python
py360convert
